container/transcriber/requirements.txt
faster-whisper>=1.0.3
diarize>=0.1.0
ffmpeg-python>=0.2.0
numpy>=1.26.0
soundfile>=0.12.0
